refactor: change `zend_is_true` to return `bool`
Previously this returned `int`. Many function actually take advantage
of the fact this returns exactly 0 or 1. For instance,
`main/streams/xp_socket.c` does:

    sockopts |= STREAM_SOCKOP_IPV6_V6ONLY_ENABLED * zend_is_true(tmpzval);

And `Zend/zend_compile.c` does:

    child = &ast->child[2 - zend_is_true(zend_ast_get_zval(ast->child[0]))];

I changed a few places trivially from `int` to `bool`, but there are
still many places such as the object handlers which return `int` that
should eventually be `bool`.

+++ Zend/zend_weakrefs.c
@@ -421,7 +421,8 @@ static int zend_weakmap_has_dimension(zend_object *object, zval *offset, int che
 	}
 
 	if (check_empty) {
-		return i_zend_is_true(zv);
+		// todo: make zend_weakmap_has_dimension return bool as well
+		return (int)i_zend_is_true(zv);
 	}
 	return Z_TYPE_P(zv) != IS_NULL;
 }
